Adding changes for Code Deploy

+++ scripts/start_server.sh
@@ -1,5 +1,8 @@
-sudo kill $(pgrep java)
-echo "Starting WinWin application"
+if pgrep -x "java" >/dev/null
+then
+    echo "Stopping WinWin application"
+	sudo kill $(pgrep java)
+	echo "Starting WinWin application"
 if [[ "$WINWIN_ENV" = "Development" ]]; then
 	nohup sudo java -Xmx768m -Dserver.port=80 -jar winwin-0.0.1-SNAPSHOT.jar /tmp 2>> /dev/null >> /dev/null &
 elif [[ "$WINWIN_ENV" = "Staging" ]]; then
@@ -7,4 +10,5 @@ elif [[ "$WINWIN_ENV" = "Staging" ]]; then
 elif [[ "$WINWIN_ENV" = "Production" ]]; then
 	nohup sudo java -Xmx3072m -Dserver.port=80 -jar winwin-0.0.1-SNAPSHOT.jar /tmp 2>> /dev/null >> /dev/null &
 fi
-echo "WinWin application started"
+echo "WinWin application started" 
+fi
